+++ realm/realm-library/src/syncIntegrationTest/java/io/realm/SyncSessionTests.java
@@ -43,30 +43,41 @@ public class SyncSessionTests extends StandardIntegrationTest {
         void onReady(SyncSession session);
     }
 
-    private SyncSession getSession() {
-        SyncUser user = UserFactory.createUniqueUser(Constants.AUTH_URL);
-        SyncConfiguration syncConfiguration = configFactory
-                .createSyncConfigurationBuilder(user, Constants.SYNC_SERVER_URL)
-                .build();
-        looperThread.closeAfterTest(Realm.getInstance(syncConfiguration));
-        return SyncManager.getSession(syncConfiguration);
+    private void getSession(SessionCallback callback) {
+        // Work-around for a race conditions happening when shutting down a Looper test and
+        // Resetting the SyncManager
+        // The problem is the `@After` block which runs as soon as the test method has completed.
+        // For integration tests this will attempt to reset the SyncManager which will fail
+        // if Realms are still open as they hold a reference to a session object.
+        // By moving this into a Looper callback we ensure that a looper test can shutdown as
+        // intended.
+        // Generally it seems that using calling `RunInLooperThread.testComplete()` in a synchronous
+        looperThread.postRunnable((Runnable) () -> {
+            SyncUser user = UserFactory.createUniqueUser(Constants.AUTH_URL);
+            SyncConfiguration syncConfiguration = configFactory
+                    .createSyncConfigurationBuilder(user, Constants.SYNC_SERVER_URL)
+                    .build();
+            looperThread.closeAfterTest(Realm.getInstance(syncConfiguration));
+            callback.onReady(SyncManager.getSession(syncConfiguration));
+        });
     }
 
     private void getActiveSession(SessionCallback callback) {
-        SyncSession session = getSession();
-        if (session.isConnected()) {
-            callback.onReady(session);
-        } else {
-            session.addConnectionChangeListener(new ConnectionListener() {
-                @Override
-                public void onChange(ConnectionState oldState, ConnectionState newState) {
-                    if (newState == ConnectionState.CONNECTED) {
-                        session.removeConnectionChangeListener(this);
-                        callback.onReady(session);
+        getSession(session -> {
+            if (session.isConnected()) {
+                callback.onReady(session);
+            } else {
+                session.addConnectionChangeListener(new ConnectionListener() {
+                    @Override
+                    public void onChange(ConnectionState oldState, ConnectionState newState) {
+                        if (newState == ConnectionState.CONNECTED) {
+                            session.removeConnectionChangeListener(this);
+                            callback.onReady(session);
+                        }
                     }
-                }
-            });
-        }
+                });
+            }
+        });
     }
 
     @Test(timeout=3000)
@@ -323,9 +334,13 @@ public class SyncSessionTests extends StandardIntegrationTest {
                         if (stringOnlies.size() == 2) {
                             Assert.assertEquals("1", stringOnlies.get(0).getChars());
                             Assert.assertEquals("2", stringOnlies.get(1).getChars());
-                            adminRealm.close();
-                            testCompleted.countDown();
-                            handlerThread.quit();
+                            handler.post(() -> {
+                                // Closing a Realm from inside a listener doesn't seem to remove the
+                                // active session reference in Object Store
+                                adminRealm.close();
+                                testCompleted.countDown();
+                                handlerThread.quit();
+                            });
                         }
                     }
                 };
@@ -340,7 +355,7 @@ public class SyncSessionTests extends StandardIntegrationTest {
             }
         });
 
-        TestHelper.awaitOrFail(testCompleted, 60);
+        TestHelper.awaitOrFail(testCompleted);
         realm.close();
     }
 
@@ -543,13 +558,15 @@ public class SyncSessionTests extends StandardIntegrationTest {
     @Test
     @RunTestInLooperThread
     public void registerConnectionListener() {
-        SyncSession session = getSession();
-        session.addConnectionChangeListener((oldState, newState) -> {
-            if (newState == ConnectionState.DISCONNECTED) {
-                looperThread.testComplete();
-            }
+        getSession(session -> {
+            session.addConnectionChangeListener((oldState, newState) -> {
+                if (newState == ConnectionState.DISCONNECTED) {
+                    // Closing a Realm inside a connection listener doesn't work: https://github.com/realm/realm-java/issues/6249
+                    looperThread.postRunnable(() -> looperThread.testComplete());
+                }
+            });
+            session.stop();
         });
-        session.stop();
     }
 
     @Test
@@ -617,11 +634,12 @@ public class SyncSessionTests extends StandardIntegrationTest {
     @Test
     @RunTestInLooperThread
     public void stop_multipleTimes() {
-        SyncSession session = getSession();
-        session.stop();
-        assertEquals(SyncSession.State.INACTIVE, session.getState());
-        session.stop();
-        assertEquals(SyncSession.State.INACTIVE, session.getState());
-        looperThread.testComplete();
+        getSession(session -> {
+            session.stop();
+            assertEquals(SyncSession.State.INACTIVE, session.getState());
+            session.stop();
+            assertEquals(SyncSession.State.INACTIVE, session.getState());
+            looperThread.testComplete();
+        });
     }
 }