+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 5.?.? (YYYY-MM-DD)
+## X.Y.Z (YYYY-MM-DD)
 
 ### Enhancements
 * None
@@ -17,9 +17,13 @@
 
 ## 5.8.0 (YYYY-MM-DD)
 
-This release contains all changes in 5.8.0-BETA1 and 5.8.0-BETA2.
+This release also contains all changes in 5.8.0-BETA1 and 5.8.0-BETA2.
 
 ### Enhancements
+* [ObjectServer] Added Subscription class available to Query-based Realms. This exposes a Subscription more directly. This class is in beta. [#6231](https://github.com/realm/realm-java/pull/6231).
+* [ObjectServer] Added `Realm.getSubscriptions()`, `Realm.getSubscriptions(String pattern)` and `Realm.getSubscription` to make it easier to find existing subscriptions. These API's are in beta. [#6231](https://github.com/realm/realm-java/pull/6231).
+* [ObjectServer] Added `RealmQuery.subscribe()` and `RealmQuery.subscribe(String name)` to subscribe immediately inside a transaction. These API's are in beta. [#6231](https://github.com/realm/realm-java/pull/6231).
+* [ObjectServer] Added support for subscribing directly inside `SyncConfiguration.initialData()`. This can be coupled with `SyncConfiguration.waitForInitialRemoteData()` in order to block a Realm from opening until the initial subscriptions are ready and have downloaded data. This API are in beta. [#6231](https://github.com/realm/realm-java/pull/6231).
 * Added support for `ImportFlag`s to `Realm.copyToRealm()` and `Realm.copyToRealmOrUpdate()`. This makes it possible to choose a mode so only fields that actually changed are written to disk. This improves notifications and Object Server performance. [#6224](https://github.com/realm/realm-java/pull/6224).
 
 ### Fixed
@@ -94,7 +98,7 @@ This release contains all changes in 5.8.0-BETA1 and 5.8.0-BETA2.
 * APIs are backwards compatible with all previous release of realm-java in the 5.x.y series.
 
 ### Internal
-* None
+* Updated to Object Store commit: 362b886628b3aefc5b7a0bc32293d794dc1d4ad5
 
 
 ## 5.7.0 (2018-09-24)